+++ test/Boolean.ts
@@ -1,5 +1,5 @@
 import { constFalse, constTrue } from 'fp-ts/function';
-import { invert, and, or, xor, allPass, anyPass } from '../src/Boolean';
+import { invert, and, or, xor, allPass, anyPass, both, either } from '../src/Boolean';
 
 describe('Boolean', () => {
     describe('invert', () => {
@@ -74,6 +74,46 @@ describe('Boolean', () => {
 
         it('returns false if all predicates fail', () => {
             expect(f([constFalse, constFalse, constFalse])(null)).toBe(false);
+        });
+    });
+
+    describe('both', () => {
+        const f = both<number>(x => x > 5)(x => x % 2 === 0);
+
+        it('fails if only first predicate fails', () => {
+            expect(f(4)).toBe(false);
+        });
+
+        it('fails if only second predicate fails', () => {
+            expect(f(7)).toBe(false);
+        });
+
+        it('fails if both predicates fail', () => {
+            expect(f(3)).toBe(false);
+        });
+
+        it('succeeds if both predicates succeed', () => {
+            expect(f(6)).toBe(true);
+        });
+    });
+
+    describe('either', () => {
+        const f = either<number>(x => x > 5)(x => x % 2 === 0);
+
+        it('fails if both predicates fail', () => {
+            expect(f(3)).toBe(false);
+        });
+
+        it('succeeds if only first predicate succeeds', () => {
+            expect(f(7)).toBe(true);
+        });
+
+        it('succeeds if only second predicate succeeds', () => {
+            expect(f(4)).toBe(true);
+        });
+
+        it('succeeds if both predicates succeed', () => {
+            expect(f(6)).toBe(true);
         });
     });
 });